started separating cudnn from dnn_lib code
